# Conflicts:
#	src/main/java/com/tp/proyecto1/controllers/VentaFormController.java
#	src/main/java/com/tp/proyecto1/controllers/ViajeFormController.java
#	src/main/java/com/tp/proyecto1/controllers/reserva/ReservaFormController.java
#	src/main/java/com/tp/proyecto1/model/viajes/Transporte.java
#	src/main/java/com/tp/proyecto1/views/reserva/ReservaForm.java

+++ src/main/java/com/tp/proyecto1/model/viajes/Viaje.java
@@ -41,11 +41,11 @@ public class Viaje {
 		this.activo = activo;
 	}
 	
-	public Double getPasajesRestantes() {
+	public Integer getPasajesRestantes() {
 		return transporte.getCapacidadRestante();
 	}
 	
-	public void restarPasajes(Double cantPasajes) {
+	public void restarPasajes(Integer cantPasajes) {
 		transporte.restarPasajes(cantPasajes);
 	}
